requirements.txt
psycopg2-binary>=2.9.0,<2.10
fastembed>=0.2.0,<0.4.0
jieba>=0.42.1,<0.43.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<2.33.0
cachetools>=5.3.0,<6.0.0
